# Conflicts:
#	core/src/com/mygdx/game/Graphics/Assets.java
#	core/src/com/mygdx/game/Levels/Level1/GameScreenLevel1.java
#	core/src/com/mygdx/game/Player/Player.java

+++ core/src/com/mygdx/game/Enemies/Cannon.java
@@ -8,6 +8,7 @@ import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.physics.box2d.*;
 import com.badlogic.gdx.utils.Array;
+import com.mygdx.game.Graphics.Assets;
 import com.mygdx.game.Levels.Level1.GameScreenLevel1;
 import com.mygdx.game.Objects.Bullet;
 import com.mygdx.game.Player.HUD;
@@ -46,7 +47,7 @@ public class Cannon extends Enemy {
         frames.add(btoom1);
         frames.add(btoom2);
         frames.add(btoom3);
-        BTOOM = new Animation(0.5f, frames);
+        BTOOM = new Animation(0.05f, frames);
         textureCannon = new Texture("Enemies/Cannon.png");
         spriteCannon = new Sprite(textureCannon);
         setBounds(getX(), getY(), 32/PPM, 32/PPM);
@@ -57,28 +58,23 @@ public class Cannon extends Enemy {
     public void update(float delta){
         stateTimer += delta;
         timer += delta;
-        if(setToDestroy && !destroyed){
+        if(setToDestroy && !destroyed && stateTimer > 0.3f){
             world.destroyBody(b2body);
             attack = false;
             setPosition(b2body.getPosition().x - getWidth()/2, b2body.getPosition().y - getHeight()/2);
-            setRegion((Sprite)BTOOM.getKeyFrame(timer, false));
             stateTimer = 0;
             destroyed = true;
             HUD.SCORE+=25;
-        }
-        else if(!destroyed) {
-            setPosition(b2body.getPosition().x - getWidth()/2, b2body.getPosition().y - getHeight()/2);
-            setRegion(spriteCannon);
         }
         if(attack){
             if (stateTimer >= shootTimer) {
                 if (!spriteCannon.isFlipX()) {
                     bullet = new Bullet(world, b2body.getPosition().x, b2body.getPosition().y, 24/PPM, 8/PPM);
-                    bullet.bulletBody.setLinearVelocity(2f, 0);
+                    bullet.bulletBody.setLinearVelocity(4f, 0);
                     cannonBullets.add(bullet);
                 } else {
                     bullet = new Bullet(world, b2body.getPosition().x, b2body.getPosition().y, -24/PPM, 8/PPM);
-                    bullet.bulletBody.setLinearVelocity(-2f, 0);
+                    bullet.bulletBody.setLinearVelocity(-4f, 0);
                     cannonBullets.add(bullet);
                 }
                 stateTimer = 0;
@@ -92,9 +88,6 @@ public class Cannon extends Enemy {
         }
 
     public void draw(Batch batch){
-        if((stateTimer < 1 || !destroyed)){
-            super.draw(batch);
-        }
         for (Bullet bullet:cannonBullets){
             if(!bullet.isDestroyed()) {
                 bullet.draw(batch);
@@ -149,6 +142,7 @@ public class Cannon extends Enemy {
     @Override
     public void deleted(){
         setToDestroy = true;
+        stateTimer = 0;
     }
 
     public void fire(){
@@ -156,5 +150,18 @@ public class Cannon extends Enemy {
     }
     public boolean isDestroyed(){
         return destroyed;
+    }
+    public Sprite babax(){
+        Sprite sprite;
+        if(setToDestroy && !destroyed){
+            sprite = (Sprite) BTOOM.getKeyFrame(stateTimer, false);
+        }
+        else if (destroyed){
+            sprite = Assets.spriteEmpty;
+        }
+        else {
+            sprite = spriteCannon;
+        }
+        return sprite;
     }
 }
