+++ core/src/com/mygdx/game/Enemies/DefendedCannon.java
@@ -11,6 +11,7 @@ import com.badlogic.gdx.physics.box2d.FixtureDef;
 import com.badlogic.gdx.physics.box2d.PolygonShape;
 import com.badlogic.gdx.physics.box2d.World;
 import com.badlogic.gdx.utils.Array;
+import com.mygdx.game.Graphics.Assets;
 import com.mygdx.game.Levels.Level1.GameScreenLevel1;
 import com.mygdx.game.Objects.Bullet;
 import com.mygdx.game.Player.HUD;
@@ -47,7 +48,7 @@ public class DefendedCannon extends Enemy {
         frames.add(btoom1);
         frames.add(btoom2);
         frames.add(btoom3);
-        BTOOM = new Animation(0.1f, frames);
+        BTOOM = new Animation(0.05f, frames);
         textureCannon = new Texture("Enemies/DefendedCannon.png");
         spriteCannon = new Sprite(textureCannon);
         setBounds(getX(), getY(), 32/PPM, 32/PPM);
@@ -58,7 +59,7 @@ public class DefendedCannon extends Enemy {
     public void update(float delta){
         stateTimer += delta;
         timer += delta;
-        if(setToDestroy && !destroyed){
+        if(setToDestroy && !destroyed && stateTimer > 0.3f){
             world.destroyBody(b2body);
             attack = false;
             setPosition(b2body.getPosition().x - getWidth()/2, b2body.getPosition().y - getHeight()/2);
@@ -67,19 +68,15 @@ public class DefendedCannon extends Enemy {
             destroyed = true;
             HUD.SCORE+=35;
         }
-        else if(!destroyed) {
-            setPosition(b2body.getPosition().x - getWidth()/2, b2body.getPosition().y - getHeight()/2);
-            setRegion(spriteCannon);
-        }
         if(attack){
             if (stateTimer >= shootTimer) {
                 if (!spriteCannon.isFlipX()) {
                     bullet = new Bullet(world, b2body.getPosition().x, b2body.getPosition().y, 24/PPM, 1/PPM);
-                    bullet.bulletBody.setLinearVelocity(2f, 0);
+                    bullet.bulletBody.setLinearVelocity(4f, 0);
                     defendedCannonBullets.add(bullet);
                 } else {
                     bullet = new Bullet(world, b2body.getPosition().x, b2body.getPosition().y, -24/PPM, 1/PPM);
-                    bullet.bulletBody.setLinearVelocity(-2f, 0);
+                    bullet.bulletBody.setLinearVelocity(-4f, 0);
                     defendedCannonBullets.add(bullet);
                 }
                 stateTimer = 0;
@@ -93,9 +90,6 @@ public class DefendedCannon extends Enemy {
     }
 
     public void draw(Batch batch){
-        if(stateTimer == 0 || !destroyed){
-            super.draw(batch);
-        }
         for (Bullet bullet:defendedCannonBullets){
             if(!bullet.isDestroyed()) {
                 bullet.draw(batch);
@@ -169,8 +163,22 @@ public class DefendedCannon extends Enemy {
     @Override
     public void deleted() {
         setToDestroy = true;
+        stateTimer = 0;
     }
     public boolean isDestroyed(){
         return destroyed;
+    }
+    public Sprite babax(){
+        Sprite sprite;
+        if(setToDestroy && !destroyed){
+            sprite = (Sprite) BTOOM.getKeyFrame(stateTimer, false);
+        }
+        else if (destroyed){
+            sprite = Assets.spriteEmpty;
+        }
+        else {
+            sprite = spriteCannon;
+        }
+        return sprite;
     }
 }
